neutrino: rework infoviewer
Make the infoviewer more "tuxbox-like". The default appearance should
only change marginally, but this provides infrastructure to:
*) disable the crypto icons (default if --disable-blinkenlights)
*) resize the infobar better to adapt to changed font sizes (not yet
 perfect, but better than before)
*) always vertically center the icons in the button bar (red green blue
 and AC3/16:9 etc)
The one thing visible immediately is that the infobar is always drawn
at the same height, regardless if the buttonbar is drawn below or not.
Originally I did it that way because a) the code was simpler and b) I
liked it better. If this is an issue, please report an I can change that.

I deliberately did not touch the display in file playback mode, this can
be done later once potential issues are shaken out.

Tis is a massive change, so I expect it to be not perfect. Please report
anything you find.

git-svn-id: file:///home/bas/coolstream_public_svn/THIRDPARTY/applications/neutrino-experimental@98 e54a6e83-5905-42d5-8d5c-058d10e6a962


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/115ba99f2ebf6f3092db287084d865da213a049a
Author: Stefan Seyfried <seife@tuxbox-git.slipkontur.de>
Date: 2009-12-20 (Sun, 20 Dec 2009)

Origin message was:
------------------
neutrino: rework infoviewer

Make the infoviewer more "tuxbox-like". The default appearance should
only change marginally, but this provides infrastructure to:
*) disable the crypto icons (default if --disable-blinkenlights)
*) resize the infobar better to adapt to changed font sizes (not yet
  perfect, but better than before)
*) always vertically center the icons in the button bar (red green blue
  and AC3/16:9 etc)
The one thing visible immediately is that the infobar is always drawn
at the same height, regardless if the buttonbar is drawn below or not.
Originally I did it that way because a) the code was simpler and b) I
liked it better. If this is an issue, please report an I can change that.

I deliberately did not touch the display in file playback mode, this can
be done later once potential issues are shaken out.

Tis is a massive change, so I expect it to be not perfect. Please report
anything you find.

git-svn-id: file:///home/bas/coolstream_public_svn/THIRDPARTY/applications/neutrino-experimental@98 e54a6e83-5905-42d5-8d5c-058d10e6a962