hdd_menu: do not allow to wipe the drive the root fs is on
If the root FS is on a HDD, disable the HDD menu for this disk.
As long as we always repartition / reformat the whole drive, this
is the best way to avoid people shooting themselves in the foot.

git-svn-id: file:///home/bas/coolstream_public_svn/THIRDPARTY/applications/neutrino-experimental@887 e54a6e83-5905-42d5-8d5c-058d10e6a962


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/b1fefb0e60227cba40d59cc2abc53938ac55d8de
Author: Stefan Seyfried <seife@tuxbox-git.slipkontur.de>
Date: 2010-11-06 (Sat, 06 Nov 2010)