neutrino scan_setup: fix double entries in satlist
It was necessary to clean satlist in 'addScanMenuAutoScanAll()',
because of multiple generated options in satlist,
This ensures that we have a clean list before we add new options.

thx SatBaby for hint

git-svn-id: file:///home/bas/coolstream_public_svn/THIRDPARTY/applications/neutrino-experimental@1497 e54a6e83-5905-42d5-8d5c-058d10e6a962


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/82be0b9f10ed9f78387fab9c04e8585a2945686b
Author: Thilo Graf <dbt@novatux.de>
Date: 2011-05-28 (Sat, 28 May 2011)

Origin message was:
------------------
*neutrino scan_setup: fix double entries in satlist

It was necessary to clean satlist in 'addScanMenuAutoScanAll()',
because of multiple generated options in satlist,
This ensures that we have a clean list before we add new options.

thx SatBaby for hint

git-svn-id: file:///home/bas/coolstream_public_svn/THIRDPARTY/applications/neutrino-experimental@1497 e54a6e83-5905-42d5-8d5c-058d10e6a962