COsdSetup: show correct option in infobar tuner icon setup item
It could be confused for single stb users to see the option
'ON', although it is not possible.
So it's better to show only a possible or real option at item, if stb is in single mode or
item is disabled,
but the real value of g_settings.infobar_show_tuner is not touched.


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/ca748c6f48a8cd2fbe21bd79356e33b5f8e025d6
Author: Thilo Graf <dbt@novatux.de>
Date: 2012-09-18 (Tue, 18 Sep 2012)

Origin message was:
------------------
*COsdSetup: show correct option in infobar tuner icon setup item

It could be confused for single stb users to see the option
'ON', although it is not possible.
So it's better to show only a possible or real option at item, if stb is in single mode or
item is disabled,
but the real value of g_settings.infobar_show_tuner is not touched.