msgbox: add rounded corners
This is not entirely correct, as the corners will be wong
if a msgbox without header or footer is used. But as the
only place where this is used is ShowMsg2UTF() which has
header and footer, it does not matter right now.


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/3ad95f30386fd7a3801dd8e9d62bac3153287280
Author: Stefan Seyfried <seife@tuxbox-git.slipkontur.de>
Date: 2013-03-24 (Sun, 24 Mar 2013)