channellist: improve add-to-bouquet and remove-new-flag workflow
When adding the current channel to a bouquet or favorites or when
removing the "new" flag from a channel, don't immediately reload the
channels but wait for closing of the channellist.
Additionally, select the next channel after one of the add-to-bouquet or
remove-new-flag actions.


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/9eda015afe24043916fbf5a15b036992a8ee70c8
Author: Stefan Seyfried <seife@tuxbox-git.slipkontur.de>
Date: 2013-04-14 (Sun, 14 Apr 2013)