CComponentsHeader: rename some members for plausible usage in subclasses
Some members are usable in subclasses, but current names could be not
suitable or not plausible with respect to the name of subclass.


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/8e46f639006faafd3026cd7ddb2d1d99fe2f6e8a
Author: Thilo Graf <dbt@novatux.de>
Date: 2013-05-27 (Mon, 27 May 2013)

Origin message was:
------------------
CComponentsHeader: rename some members for plausible usage in subclasses

Some members are usable in subclasses, but current names could be not
suitable or not plausible with respect to the name of subclass.