configure: add libbluray support into configure
It may happen that the build process aborts if this lib is not
expilzit specified, although this lib is not used directly
by neutrino.
It seems to be dependent of build environment, and should be understand
as a prevention.


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/f2bfa22159f997a50b7994900d0f028ba24e7f51
Author: Thilo Graf <dbt@novatux.de>
Date: 2014-02-23 (Sun, 23 Feb 2014)