CMenuWidget: add member to suppress paint of details line
In some situations is it more senseful to suppress painting of details line.
Required if menu window is painted over another window.
The hide() functionality of detailsline paints only an empty background on screen so we have
blank parts on screen. That looks bad. For example, this is to observe
on context menu in channellist.

We should see this as a workaround. It is not certain whether the
details line is really needed in the future.


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/793230e9a7b521b07bc9ac55ce9dc4c951b16980
Author: Thilo Graf <dbt@novatux.de>
Date: 2014-06-09 (Mon, 09 Jun 2014)