CComponentsFooter: disable gradient for footers
This is not really completly implemented for footer at the momemnt, because
button objects are still not prepared or tuned for that.

If it is enabled, currently it could occur unintentionally background behavior
eg.: flickereffects or background display errors in button objects.

NOTE: anyway users can enable gradient in theme files and could override this setting.
Now here it is globally disabled for footers. If it works right,
it will be enabled in future.


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/427b65813d9f0df6b05b18407ae70ca966ddd482
Author: Thilo Graf <dbt@novatux.de>
Date: 2016-11-27 (Sun, 27 Nov 2016)

Origin message was:
------------------
CComponentsFooter: disable gradient for footers

This is not really completly implemented for footer at the momemnt, because
button objects are still not prepared or tuned for that.

If it is enabled, currently it could occur unintentionally background behavior
eg.: flickereffects or background display errors in button objects.

NOTE: anyway users can enable gradient in theme files and could override this setting.
Now here it is globally disabled for footers. If it works right,
it will be enabled in future.