[RemoteTimers] add possibility to transfer all timers at once via webif
Use:
http://boxip/control/sendalltimers?ip=10.1.1.xxx&force=1
http://boxip/control/sendalltimers?name=record%20box&force=true

if parameter name is given, it has to be in timerlist as remote box.
if parameter ip and name is given, ip is used.
if parameter force is set, all checks are disabled, and the local timer
will be deleted even the transfer was not successful.


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/7031e99573959206d0cd887f04c2ebe8d0b8d408
Author: TangoCash <eric@loxat.de>
Date: 2017-06-15 (Thu, 15 Jun 2017)