CComponentsHeader: rework and simplified logo handling
This should fix some issues with possible overlapps
of logos with title. With some Conditions with text
alignment on the right side needs a fix.
There is an issue with size and position of text object, but is stll not
used at the moment. For all other conditions of logo and text alignments now
it should be not a problem (I hope) and have more code clarity.


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/356719cf3a771cb2eb667969e7072b97a6501e0b
Author: Thilo Graf <dbt@novatux.de>
Date: 2017-08-02 (Wed, 02 Aug 2017)