neutrino: always write /tmp/.timer ...
otherwise the vfd stays blank in deepstandby, if leds are configured off
and no timer is set. HD2's vfd urgently needs the fp.brightness value
which is not written to frontpanel in this case.


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/51ce12c3ee4133cda671360871da874300ac61b2
Author: vanhofen <vanhofen@gmx.de>
Date: 2017-11-20 (Mon, 20 Nov 2017)

Origin message was:
------------------
- neutrino: always write /tmp/.timer ...

otherwise the vfd stays blank in deepstandby, if leds are configured off
and no timer is set. HD2's vfd urgently needs the fp.brightness value
which is not written to frontpanel in this case.