icons: rename mute.png => muted.png; ...
This icon isn't a button. It signalizes the mute-status. So i renamed it to get the name free for the button icon.


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/db25f2c0dcabcadc98b76b7a0daa1467e11ea206
Author: vanhofen <vanhofen@gmx.de>
Date: 2018-01-04 (Thu, 04 Jan 2018)

Origin message was:
------------------
- icons: rename mute.png => muted.png; ...

This icon isn't a button. It signalizes the mute-status. So i renamed it to get the name free for the button icon.