neutrino: fix infobar display with disabled timeout
When the infobar timeout is disabled, the infobar was not shown for all
"automatic" events like "start" and "zap".
The purpose of disabled infobar timeout is to show the infobar all the
time (e.g. in radio mode), so this does not make sense. Fix it.


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/9cd7e8972cf2adccadeec25a872281e15116836f
Author: Stefan Seyfried <seife@tuxbox-git.slipkontur.de>
Date: 2018-02-04 (Sun, 04 Feb 2018)