menu: fix return value of getSelectedLine() for RC_timeout and RC_setup
This fixes the bahaviour of StartPosSelectionMenu in moviebrowser.
When this menu was closed by its timeout or RC_setup was pressed
the movie starts anyway.


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/70cbc65483598c1a69d57b0dbc179e1864598405
Author: vanhofen <vanhofen@gmx.de>
Date: 2018-07-15 (Sun, 15 Jul 2018)

Origin message was:
------------------
- menu: fix return value of getSelectedLine() for RC_timeout and RC_setup

This fixes the bahaviour of StartPosSelectionMenu in moviebrowser.
When this menu was closed by its timeout or RC_setup was pressed
the movie starts anyway.