Remove dead code from CRCInput::getMsg_us()
The two events in question are posted in CInfoViewer::getEPG()
and they're handled in CRemoteControl::handleMsg().

They are written to the high/low priority pipes only, so there should
be no way that they end up in the event socket, which is what the code
path in question here deals with. Apart from that, the type the buffer
was casted to was completely wrong, so if this code was ever executed,
it would have faulted.


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/58966baac8b49c2dc6fe5d211b01871470d08da4
Author: Hendi <hendi48@freenet.de>
Date: 2023-01-23 (Mon, 23 Jan 2023)