system/helpers: Add new Parameter 'start_pos' to str_replace()
This allows flexibility to specify a starting position from where to search
for the first occurrence of the search string. By default,
the position is set to 0 and the function will behave as before.
However, when a position is provided,
the function will replace only the first occurrence of
the search string after that position.


Origin commit data
------------------
Commit: https://github.com/neutrino-images/ni-neutrino/commit/912f256276433fb04bdc2caed3704c8f973808ae
Author: Thilo Graf <dbt@novatux.de>
Date: 2023-03-07 (Tue, 07 Mar 2023)